not-to-be-proud-of fix of embedded interpreter init
We're supposed to pass command-line arguments to the R function
initializing the interpreter. Problem: now they are now immutable
strings, which are now const char* on the C side. But the R code
provides no guarantee on this and takes char*... This is detected as
an error on some gcc versions, and is indeed not safe.

The correct solution would be to copy the strings but then I have no
hook to free the corresponding memory, which would result in a memory
leak. While it's not particularly serious, I chose for now to keep the
statu quo: I interpret args as Bytes on the C side.