Fix configurator output for empty cflags list
I was having some trouble upgrading the version of ocaml-r in nixpkgs to 0.4.0, and noticed that gcc was being passed a spurious empty string argument.  I believe the recent change to this file may have resulted in empty flag sets being output to the sexp file as a single quoted empty string instead of an empty list as intended.